Address additional Copilot review feedback
- vector.go: Fix type mismatch in float32ToFloat16 bitwise OR (cast all operands to uint32)
- tds.go: Update TODO comment to reflect float32 uses binary, float16 uses JSON
- doc/how-to-use-vectors.md: Clarify float16 is sent as JSON but received as binary
- vector_db_test.go: Remove broken t.Cleanup() (was tied to first test, didn't work with sync.Once)
  - Drop-before-create pattern handles cleanup on next test run instead

+++ vector.go
@@ -503,7 +503,7 @@ func float32ToFloat16(value float32) uint16 {
 		}
 	}
 
-	return uint16((sign << 15) | (uint32(halfExponent) << 10) | mant)
+	return uint16((uint32(sign) << 15) | (uint32(halfExponent) << 10) | uint32(mant))
 }
 
 // float16ToFloat32 converts a float16 (IEEE 754 half-precision) value to float32.
